refactor: change the structure of the project
